notebookLM-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
supabase==2.0.2
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
fastembed==0.2.0
pypdf2==3.0.1
pymupdf==1.23.8
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
